+++ dgisim/src/character/character.py
@@ -34,6 +34,7 @@ __all__ = [
     "AratakiItto",
     "Bennett",
     "ElectroHypostasis",
+    "FatuiPyroAgent",
     "KaedeharaKazuha",
     "Kaeya",
     "Keqing",
@@ -716,6 +717,80 @@ class ElectroHypostasis(Character):
             energy=0,
             max_energy=2,
             hiddens=stts.Statuses((stt.ElectroCrystalCoreHiddenStatus(),)),
+            equipments=stts.EquipmentStatuses(()),
+            statuses=stts.Statuses(()),
+            elemental_aura=ElementalAura.from_default(),
+        )
+
+
+class FatuiPyroAgent(Character):
+    _ELEMENT = Element.PYRO
+    _WEAPON_TYPE = WeaponType.NONE
+    _TALENT_STATUS = stt.PaidInFullStatus
+    _FACTIONS = frozenset((Faction.FATUI,))
+
+    _NORMAL_ATTACK_COST = AbstractDices({
+        Element.PYRO: 1,
+        Element.ANY: 2,
+    })
+    _ELEMENTAL_SKILL1_COST = AbstractDices({
+        Element.PYRO: 3,
+    })
+    _ELEMENTAL_BURST_COST = AbstractDices({
+        Element.PYRO: 3,
+    })
+
+    @override
+    def _normal_attack(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return normal_attack_template(
+            game_state=game_state,
+            source=source,
+            element=Element.PHYSICAL,
+            damage=2,
+        )
+
+    @override
+    def _elemental_skill1(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.PYRO,
+                damage=1,
+                damage_type=DamageType(elemental_skill=True),
+            ),
+            eft.UpdateCharacterStatusEffect(
+                target=source,
+                status=stt.StealthStatus(usages=3 if self.talent_equiped() else 2),
+            ),
+        )
+
+    @override
+    def _elemental_burst(self, game_state: GameState, source: StaticTarget) -> tuple[eft.Effect, ...]:
+        return (
+            eft.EnergyDrainEffect(
+                target=source,
+                drain=self.get_max_energy(),
+            ),
+            eft.ReferredDamageEffect(
+                source=source,
+                target=DynamicCharacterTarget.OPPO_ACTIVE,
+                element=Element.PYRO,
+                damage=5,
+                damage_type=DamageType(elemental_burst=True),
+            ),
+        )
+
+    @classmethod
+    def from_default(cls, id: int = -1) -> Self:
+        return cls(
+            id=id,
+            alive=True,
+            hp=10,
+            max_hp=10,
+            energy=0,
+            max_energy=2,
+            hiddens=stts.Statuses((stt.StealthMasterStatus(),)),
             equipments=stts.EquipmentStatuses(()),
             statuses=stts.Statuses(()),
             elemental_aura=ElementalAura.from_default(),